raneys: Work on pf of prop 4; add pandoc-crossref.
I'm not super-happy with pandoc-crossref because it was a bear to
install, it only works on display equations that are their own
paragraph, and it doesn't right-justify the equation numbers in html.
I'm tempted to write my own version of this, but - let's be honest - I
don't really need to and it's probably not the best use of my time right
now.