raneys: start pf of lemma 9, add g_k outline img.
I found it challenging to get the image close to a quality level I'm
ok with. I wish I knew how to set it up entirely as a vector-based
image. I also wish I knew how to exclude the figure caption in the pdf
output; in that case I'd have excluded the caption entirely in both
output formats.