self_repl: Make 1st fn addition image.
This includes added support for SVG-based coloring in of the area under a
function.

I'm not happy with the current colors used. Also, ideally, it would be easy to
change these colors, but for now it's a bit more work as there is a light
version for the fill and a dark version for the stroke; both are currently
hard-coded values.