self_repl: Standardize the image style and colors.
I wish it were easier for me to update all of them at once. I might be able to
get closer to that by using a stylesheet for the image colors and assigning the
relevant svg elements to classes. In order to make the end-to-end system easier,
that would also require using svg images in html. Even then, though, it is still
a best practice to have a png file as a fallback.