self_repl: Fix spacing around internal refs.
I found this tricky to figure out. This comment on a bug was the first big clue:

https://github.com/lierdakil/pandoc-crossref/issues/52

followed by a fair amount of trial and error. I think I was stuck for a while
not realizing that the string "$$i$$" was telling bash to replace $$ with the
pid of the current script.