self_repl: update nonplateau setup image.
I had a lot of trouble getting text in a vector image to look good in both html
and pdf. One of my goals was to arrive at a way to build images that preserved
vectors as much as possible. Hence I'd like vector-based pdf versions of the
images that will become embedded within the final pdf. My original toolchain,
svg crowbar -> cairosvg, did not correctly handle the text fonts that I put into
my svg. In particular, cairosvg dropped some critical font information.

The solution I eventually came up with was to manually add the text later using
Inkscape, and to save as a pdf from there. In one sense, this is sad in that
adding text requires an entirely new step, and can't be done programmatically as
far as I know. Looking on the bright side, though, in the cases I anticipate,
adding text via Inkscape will actually be easier than adding it via JavaScript.
And I have achieved my goal of matching html/pdf with a pure-vector pdf.