self_repl: clarify eqn (3), add kindle notes.
I'm happy to see how pleasant it can be to read a paper from this system on a
kindle. I'm thinking of making a kindle directory that includes a bash script
which copies over files and makes minor adjustments and then compiles everything
into a nice kindle-friendly pdf. These are the changes I have in mind:

* The latex updates mentioned in kindle_notes.md.
* Increase the contrast on the image colors to look good in black and white.
* Increase the white divider line thickness in images.
* Minor math tweaks such as for some overlines and the \zerotwo macro.