self_repl: add \latexonlyrule command.
Enabling commands that depend on platform required me to add template.tex which
is, for now, a very slight modification of pandoc's default latex template.

My practical goal here was to improve the spacing within a math array
environment. There may have been a more general or best-practicey way to do
this, but this works and I got it done reasonably quickly.