cs229: Work on problem set 1.
I am ready to finish problem 1, but I'm frustrated by trying to use octave on
macOS. The graphing functions are ancient, look bad, and are difficult to use.
Because of that, I'm thinking of switching over to numpy from here on out. I may
go back to matlab/octave for problems that don't involve any graphing, although
even then the ability to graph is nice for debugging and sanity checking
purposes.