img12: Messy code but I believe bug fixed.
This time it was reproducible. You had to click on a face, don't move
the mouse at all, and then mouse up when another face had rotated in.

In the future I plan to make a click like this pause the polytope so
that it feels like a click instantly puts us in dragging mode, which I
think is the more intuitive behavior.