circle_test1: Find a basis for the circle.
This is a work in progress. The plan is to use the basis to find two
pairs of corner points for the square that surrounds the circle, and
such that that two near points have the same z coordinates, as do the
far two points. Once we have that, we can easily find the center (in
screen coordinates) of the ellipse, as well as one axis (probably the
minor one). The plan is to then use the tangent point within the
perspective-transformed square to find the major ellipse axis, which
completely determines the screen ellipse.