img13: Better support turning on space.js anim'n.
Note: If we turn on animation (eg rotation), there is still an issue in
that the dot sizes are not quite right anymore. They are memorized at
the start of the animation, and the current dot-highlighting system
assumes that the base radius will stay the same (which is true when
there's no rotation, but is false otherwise).