Begin draw2.
My plan is for draw2 to eventually be renamed as draw.js and to replace
that interface completely.

The motivation for this is my expectation that canvas rendering will be
smoother (better framerate) than svg rendering in some cases, and I'd
like to keep my current interface (with some changes) but to let that
render to a canvas instead of as an svg.

Some notes for future me:
* Rendering efficiency is important. Try hard to keep things smooth.
* Switch from {x, y} coordinates to [x, y].
* I believe I'll need to add a render() call, or something equivalent,
  to mark the completion of changes to a canvas element. Is there a
  clever way to avoid that necessity?
* I may want to add support to see if a point is inside an element so
  that canvas rendering can still be responsive to geometrically-aware
  mouse events.