draw2/img12:work-in-progress: mouse/out/over.
I realized that the browser has a lot of structure around these events.
It seems that only a single element can have the mouse over it at a
time, and that the previous element always receives a mouseout event
*before* the new element receives a mouseover event. I think I will take
a minor shortcut and assume that the user is providing me with a
mutually exclusive set of polygons to receive this event (although these
polygons need not fill the canvas).

A couple next steps:
* Pull out this (event-dispatching) logic into a little js file. It
  doesn't fully fit into draw2, conceptually.
* React appropriately to mouseout on the entire canvas.
* React appropriately when a polygon moves. This is somewhat subtle in
  that we could have polygons temporarily overlap as they are moved one
  by one. Because of that case, I think it's appropriate to choose an
  arbitrary-but-consistent ordering for them, and to send events with
  the guarantees above (one-over-at-a-time; mouseout-sent-first).
* Actually render text in a canvas.
* Correct the sizing of the label background rounded rects.