draw2: Updates to img 8,9.
I modified the calling convention for some fns in perm2.js to work with
a final `params` object instead of passing in multiple optional values.
One thing this helps with is passing through parameters downstream.

I enabled G_8 on img08; however, the rendering seems to partially break
down at this point (within Chrome), so it doesn't add much value. On the
positive side, at least Chrome with canvas is able to put some kind of
somewhat meaningful image in just a few seconds at this scale. Some
sizes of the rendering look correct, but at low resolutions we get some
kind of moire effect, and at high resolutions it appears to me to be
buggy (again, on the rendering side, not in the js, I believe).