Work-in-progress: Updating img13 for draw2.
Currently it still works completely in svg, and partially works in
canvas, but the 3d circle is not yet rendering.

I will have to do a little work to fix that. I think the easiest way to
accomplish this is to create a Path2D instance using the path
description I've already figured out in space2.renderCircle().