Fix for Safari.
There was a Safari error when I was using a private method naming
convention (starting method names with #). I'm a bit confused by this,
since mdn says Safari supports it, but for now this is a simple fix that
appears to work. (With the drawback that the "private" methods are no
long enforced as private; not a big deal in my opinion.)