img22: Tweak layout of the flattened perm'hedron.
At this point I feel this image (img22) is in a fairly good place.

If I were to work on this more, I think the next thing I'd do is to
adjust the color scheme a little.

I also am aware that the background white behind a label is not
currently rendered nicely in Chrome (and I think in Safari?), but this
is relatively small potatoes.