Start section 2; figure out cycle notation.
I was hoping to define a custom command to help me write out cycle
notation, but I think it's difficult to do this in terms of pandoc
converting markdown to html. For now, it's really not too hard to just
type out \; between numbers in cycle notation (instead of a space).
However, part of my brain is curious thinking about how one might, in
general, support more nuanced macros that could work for different types
of output.