Ensure all lines are drawn.
At this point, the order is still arbitrary.

I'm fairly certain that I can simplify the dependency-based code. For
example, right now each time something is drawn, I iterate over
everything that might depend on it, and then call a method on each of
those. I suspect I can instead call a function on the just-drawn object
that incorporates that loop for me.