Work on face-face overlap check.
In particular, this code finds if one vertex of one face is visually
within the other face. From this, we can then compute the z values at
that point, and determine who is in front of whom.

That will not be the end of the face-face comparison, though. It's
possible for one face to be in front of the other without any vertices
being in the overlap. To handle that case, I plan to also run an
edge-edge overlap check.

I have a proof somewhere that these two checks handle all cases. If I
remember correctly, I concluded that this is akin to the polygon
version of the Jordan curve theorem; any polygon P splits the plane
into two connected regions. So if two points on the boundary of another
polygon Q (not necessarily vertices of Q) cross the original polygon P,
then some point of Q must intersect P. Otherwise, the set R^2\P would be
connected, contradicting that it consists of two disjoint connected
regions.