Edge-edge comparison is better.
I took a screenshot that shows a couple problems, though:
 (i) Some lines are still drawn a bit out of order. I believe what's
     happening is that the sorting is correct, but we don't always obey
     the sorted order at render time.
(ii) I also noticed that I re-render some edges.

The problem appears to be around dependency-triggered rendering. I'd
like to think about how that overall system works, and see if I can come
up with a solution that both single-renders everything, and that
respects the sorted order within the constraint of the dependencies.