sort.js: Drop the `opts` value.
This is replaced by using object properties of the Sorter class.

Also, I had just thought that I had a long-lasting bug in counting
numCmpCalls, but I think that doubt was incorrect, and in fact I think I
had a short-term bug that *over*-counted them, and I think that bug is
now fixed. To be clear, I think my metrics have been all correct except
for a short-term blip which has been fixed as of this commit.