Work-in-progress: Fixing sort.
Not all tests are passing right now. I paused work on this for a little
bit, and am trying to get back into it. If I remember correctly, I had
realized that tracking a tree alone is not enough info, and that I want
to track a full DAG. I believe I'm in the process of upgrading the
algorithm in that sense.