Slight code simplification & fix.
The issue being addressed here is that the inner loop in the sort
algorithm would in some cases re-check roots that we'd already covered;
this could happen since I was previously using two separate copies (one
was an array, the other a set) of the list of roots remaining to sort. I
got rid of the array since I believe the time complexity is better this
way.