Add cycle warning and try to recover from them.
I'm not sure the cycle recovery 100% works because an example scene with
intersections still froze. In the long run, it would be nice to not
freeze even with intersections; just render poorly in that case. Best
outcome is that the artifacts are only around the offending shapes and
others do what they should, which I don't think is currently the case.